requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.0.7
djangorestframework==3.13.1
idna==3.3
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
dj-database-url==1.0.0
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.9.3
whitenoise==6.2.0
django-cors-headers==3.13.0
python-dateutil==2.8.2